% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{meta_data}
\alias{meta_data}
\title{covariates information for Karlsson et al.}
\format{A data frame containing 145 rows and 12 columns, the first column indicates the condition we are interested in.
\describe{
\item{study_condition}{1}
\item{age}{2}
\item{number_reads}{2}
\item{triglycerides}{2}
\item{hba1c}{2}
\item{ldl}{2}
\item{c_peptide}{2}
\item{cholesterol}{2}
\item{glucose}{2}
\item{adiponectin}{2}
\item{hscrp}{2}
\item{leptin}{2}
}}
\usage{
meta_data
}
\description{
The covariate information for 145 samples (subjects) with 12 features
}
\keyword{datasets}
